refactor: 구현체 내부에서 실제 로직이 노출되도록 리팩토링
